requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
fsspec==2025.2.0
huggingface-hub==0.29.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pypdfium2==4.30.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==75.8.0
surya-ocr==0.13.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
